Use `f32` instead of `i32` for lengths
This allows users to use logical coordinates instead of physical ones.

+++ examples/editor-test/src/main.rs
@@ -31,7 +31,7 @@ fn redraw(window: &mut Window, editor: &mut Editor<'_>, swash_cache: &mut SwashC
 fn main() {
     env_logger::init();
 
-    let display_scale = 1;
+    let display_scale = 1.0;
     let font_system = FontSystem::new();
 
     let mut window = Window::new_flags(
@@ -45,17 +45,17 @@ fn main() {
     .unwrap();
 
     let font_sizes = [
-        Metrics::new(10, 14).scale(display_scale), // Caption
-        Metrics::new(14, 20).scale(display_scale), // Body
-        Metrics::new(20, 28).scale(display_scale), // Title 4
-        Metrics::new(24, 32).scale(display_scale), // Title 3
-        Metrics::new(28, 36).scale(display_scale), // Title 2
-        Metrics::new(32, 44).scale(display_scale), // Title 1
+        Metrics::new(10.0, 14.0).scale(display_scale), // Caption
+        Metrics::new(14.0, 20.0).scale(display_scale), // Body
+        Metrics::new(20.0, 28.0).scale(display_scale), // Title 4
+        Metrics::new(24.0, 32.0).scale(display_scale), // Title 3
+        Metrics::new(28.0, 36.0).scale(display_scale), // Title 2
+        Metrics::new(32.0, 44.0).scale(display_scale), // Title 1
     ];
     let font_size_default = 1; // Body
 
     let mut buffer = Buffer::new(&font_system, font_sizes[font_size_default]);
-    buffer.set_size(window.width() as i32, window.height() as i32);
+    buffer.set_size(window.width() as f32, window.height() as f32);
 
     let mut editor = Editor::new(buffer);
